LayoutTests/inspector-obsolete/http-tests/search/search-in-script.html

<html>
<head>
<script src="../inspector-test.js"></script>
<script src="../resources-test.js"></script>
<script src="../debugger-test.js"></script>
<script src="search-test.js"></script>
<script>
function test()
{
    InspectorTest.startDebuggerTest(step1);
    var script;

    function step1()
    {
        InspectorTest.runAfterResourcesAreFinished(["search.js"], step2);
    }

    function step2()
    {
        InspectorTest.showScriptSource("search.js", step3);
    }

    function step3()
    {
        var url = "http://127.0.0.1:8000/inspector/search/resources/search.js";
        var scripts = InspectorTest.queryScripts(function(s) { return s.sourceURL === url; });
        script = scripts[0];
        InspectorTest.addResult(script.sourceURL);

        // This file should not match search query.
        var text = "searchTest" + "UniqueString";
        script.searchInContent(text, false, false, step4);
    }

    function step4(searchMatches)
    {
        InspectorTest.dumpSearchMatches(searchMatches);

        // This file should not match search query.
        var text = "searchTest" + "UniqueString";
        script.searchInContent(text, true, false, step5);
    }

    function step5(searchMatches)
    {
        InspectorTest.dumpSearchMatches(searchMatches);

        var text = "[a-z]earchTestUniqueString";
        script.searchInContent(text, false, true, step6);
    }

    function step6(searchMatches)
    {
        InspectorTest.dumpSearchMatches(searchMatches);

        var text = "[a-z]earchTestUniqueString";
        script.searchInContent(text, true, true, step7);
    }

    function step7(searchMatches)
    {
        InspectorTest.dumpSearchMatches(searchMatches);

        InspectorTest.completeDebuggerTest();
    }
}
</script>
</head>
<body>
<p>Tests script search in inspector debugger agent.</p>
<a href="https://bugs.webkit.org/show_bug.cgi?id=69015">Bug 69015</a>

<iframe src="resources/search.html" onload="runTest()">
</body>
</html>
